fix: 修复 SD/MM 模块编译错误
- 补充 Repository 缺失的分页查询方法
- 修复 MmInventoryController 类型转换问题
- 修复 WorkflowExpressionParserTest 缺少 List 导入

+++ backend-modulith/src/main/java/com/nexterp/business/supply/controller/MmInventoryController.java
@@ -66,7 +66,7 @@ public class MmInventoryController {
         if (materialId != null) {
             List<StockDTO> list = inventoryService.listStockByMaterial(materialId, tenantId);
             PageResult<StockDTO> result = PageResult.<StockDTO>builder()
-                    .records(list).total(list.size()).current(1).size(list.size()).build();
+                    .records(list).total((long) list.size()).current(1).size(list.size()).build();
             return Result.success(result);
         }
         return Result.success(PageResult.<StockDTO>builder()
